Refining algorithm to work with duplicated data and take time intersections

+++ Employees.Core/Employees/EmployeesProjectDaysModel.cs
@@ -8,6 +8,6 @@
 
         public uint ProjectID { get; set; }
 
-        public ulong DaysOfCoworking { get; set; }
+        public int DaysOfCoworking { get; set; }
     }
 }
